+++ scripts/mruby/linux.rb
@@ -5,5 +5,6 @@ MRuby::Build.new do |conf|
   conf.gem github: 'iij/mruby-dir'
   conf.gem github: 'iij/mruby-iijson'
   conf.gem github: 'iij/mruby-mtest'
+  conf.gem github: 'katzer/mruby-tiny-opt-parser'
   conf.gem github: 'hellrok/mruby-require'
 end